Refactor including test folders
